More unit test fixes
